+++ NEWS.md
@@ -1,6 +1,7 @@
 0.0.7.001
 ===================
-- Fix bug in `trim_osmdata` so that all sf attributes are reinstated
+- Fix bug in `trim_osmdata` so that all sf attributes are reinstated, and also
+  issue message that sf-preload is necessary for this function
 
 0.0.7
 ===================